Move HTTP mock helpers to datadog_checks_dev
Move HTTPResponseMock and RequestWrapperMock from
datadog_checks.base.utils.http_mock to datadog_checks.dev.http so
they are not shipped with the base check. Add http_response_mock
and request_wrapper_mock pytest fixtures in the dev plugin. Update
base tests and docs to use datadog_checks.dev.http.
