+++ datadog_checks_base/tests/base/utils/http/test_http_mock.py
@@ -7,7 +7,7 @@ import pytest
 
 from datadog_checks.base import AgentCheck
 from datadog_checks.base.utils.http import RequestsWrapper
-from datadog_checks.base.utils.http_mock import HTTPResponseMock, RequestWrapperMock
+from datadog_checks.dev.http import HTTPResponseMock, RequestWrapperMock
 from datadog_checks.base.utils.http_protocol import (
     HTTPClientProtocol,
     HTTPResponseProtocol,
